third_party/blink/web_tests/fast/dom/shadow/alternate-stylesheets-expected.txt

Title attribute should be ignored for style elements in shadow trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(root.querySelector('div')).color is "rgb(0, 128, 0)"
PASS getComputedStyle(alternate).color is "rgb(0, 128, 0)"
PASS host.shadowRoot.styleSheets[0].title is null
PASS successfullyParsed is true

TEST COMPLETE

Should be green
